+++ tests/integration/cases/call_variable_form_no_args/Ada_call@ada_2022.adb
@@ -1,0 +1,7 @@
+with A_Stub; use A_Stub;
+procedure Main is
+    function Make_Widget return A_Val is (ANull);
+    my_data : A_Val := Make_Widget;
+begin
+    null;
+end Main;