+++ tests/integration/cases/call_variable_form_no_args/CSharp_call@v10.cs
@@ -1,0 +1,7 @@
+using System;
+class Check {
+static object make_widget() => null;
+    public static void Main() {
+var my_data = make_widget();
+    }
+}