+++ tests/integration/cases/call_variable_form_no_args/Elixir_call@v1_14.ex
@@ -1,0 +1,7 @@
+defmodule Check do
+  def make_widget(), do: nil
+  def x do
+    my_data = make_widget()
+    _ = my_data
+  end
+end