+++ tests/integration/cases/call_variable_form_no_args/FSharp_call@v8.fs
@@ -1,0 +1,6 @@
+module Main
+
+let make_widget () : obj = null
+type Val =
+    | FList of Val list
+let my_data = make_widget()