+++ tests/integration/cases/call_variable_form_no_args/OCaml_call@v5.ml
@@ -1,0 +1,8 @@
+module Check = struct
+
+let make_widget _ = ()
+type val_t =
+  | OList of val_t list
+let my_data = make_widget()
+
+end